[Enhancement] Support VARCHAR up to 2 GiB minus 1 KiB (#59167)
* [Enhancement] Support VARCHAR up to 2 GiB minus 1 KiB

Raise the OLAP VARCHAR limit to 2 GiB minus 1 KiB across FE and BE, validate the configured upper bound, and use it for inferred and materialized schemas and statistics storage. Preserve external-catalog VARCHAR compatibility lengths and the 1 MiB SQL VARBINARY limit.

Harden execution, transport, and storage for large strings with overflow-safe serialized keys, bounded reservations, lower Thrift peak memory, and MySQL batch splitting against soft, Thrift, and BRPC limits. Reject oversized single rows, align CSV, RPC, and MySQL size handling, disable zonemaps for values over 1 MiB, and fall back from dictionary to plain encoding based on actual values while preserving mixed-page reads and global-dictionary validity.

Filter VARCHAR values over 1 MiB from histogram MCV, bucket, and HLL aggregation. Apply runtime filters to every VARCHAR column while keeping grouping, sorting, and aggregation on the original column, trading accuracy on oversized values for collection stability.

Update English, Chinese, and Japanese documentation for VARCHAR limits, inferred schemas, VARBINARY compatibility, and large-value storage behavior.


* style: auto-fix clang-format issues

---------

+++ be/src/column/mysql_row_buffer.h
@@ -113,7 +113,7 @@ public:
 
     void separator(char c);
 
-    int length() const { return _data.size(); }
+    size_t length() const { return _data.size(); }
 
     // move content into |dst| and clear this buffer.
     void move_content(std::string* dst) {
